major all working

+++ src/components/CardItem.js
@@ -1,11 +1,9 @@
 import React from 'react';
 
 const CardItem = (props) => {
-  
-
   function handledelete(){
-    // document.getElementById("empname").innerHTML = `Removing this employee named "${props.name}" makes him/her removed completely from db!!`
-
+    let confi= window.confirm("Are you sure!! You want to delete Employee named "+props.name)
+    if(confi){
       try {
         fetch(`http://localhost:9000/delete/${props.value}`,{
           method:"DELETE"
@@ -13,8 +11,6 @@ const CardItem = (props) => {
         .then((res)=>{
           let updatedEmployees;
           if(res.ok){
-            console.log(props.name)
-            // console.log("confirmed delete")
             updatedEmployees = props.employees.filter(emp=>emp._id!==props.value)
           }
           props.setEmployees(updatedEmployees);
@@ -22,12 +18,47 @@ const CardItem = (props) => {
       } catch (error) {
         console.log(error)
       }
-      
     }
-  
-  function handleEdit(){
-    
-  }
+    }
+
+    async function handleEdit(){
+      let response = await fetch(`http://localhost:9000/employee/${props.value}`)
+      let data = await response.json();
+      let name = document.getElementById("employee-name")
+      let photo = document.getElementById("photo")
+      let role = document.getElementById("role")
+      let id = document.getElementById("id")
+      id.value = data._id
+      name.value = data.name
+      photo.value = data.photo
+      role.value = data.role
+    }
+
+      async function handleeditsubmit(){
+        let response = await fetch(`http://localhost:9000/employee/${props.value}`)
+        let data = await response.json();
+        let name = document.getElementById("employee-name").value
+        let photo = document.getElementById("photo").value
+        let role = document.getElementById("role").value
+        let id = document.getElementById("id").value
+        let eresponse =  await fetch(`http://localhost:9000/edit/${id}`,{
+          method:"PUT",
+          headers:{
+            'Content-Type': 'application/json',
+          }, 
+          body:JSON.stringify({
+            name:name,
+            photo:photo,
+            role:role
+          })
+        })
+        const updatedEmployees = props.employees.map((employee) =>
+          employee._id === id
+            ? { ...employee, name: name, photo: photo, role: role }
+            : employee
+        );
+        props.setEmployees(updatedEmployees);
+      }
 
   return (
     <>
@@ -45,19 +76,18 @@ const CardItem = (props) => {
           gap:"5px",
           maxWidth:"8rem"
         }}>
-          <button className='btn btn-warning' 
-          onClick={handleEdit}>Edit</button>
-          <button className='btn btn-danger' onClick={handledelete} 
+          <button className='btn btn-warning' onClick={handleEdit} data-bs-toggle="modal" data-bs-target="#exampleModalEditEmp">Edit</button>
+          <button className='btn btn-danger' onClick={handledelete}
           >Delete</button>
         </div>
       </div>
 {/* modal view */}
-{/* data-bs-toggle="modal" data-bs-target="#deleteEmp"  that needed to be added while activating modal*/}
-<div className="modal fade" id="deleteEmp" data-bs-backdrop="static" data-bs-keyboard="false" tabIndex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+{/*   that needed to be added while activating modal data-bs-toggle="modal" data-bs-target="#deleteEmp"*/}
+{/* <div className="modal fade" id="deleteEmp" data-bs-backdrop="static" data-bs-keyboard="false" tabIndex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
   <div className="modal-dialog">
     <div className="modal-content">
       <div className="modal-header">
-        <h1 className="modal-title fs-5"      id="staticBackdropLabel">Delete This Employee</h1>
+        <h1 className="modal-title fs-5" id="staticBackdropLabel">Delete This Employee</h1>
         <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
       </div>
       <div className="modal-body" id='empname'>
@@ -69,8 +99,74 @@ const CardItem = (props) => {
       </div>
     </div>
   </div>
-</div>
+</div> */}
       
+
+      {/* edit modal view */}
+      <div
+        className='modal fade'
+        id='exampleModalEditEmp'
+        tabIndex='-1'
+        aria-labelledby='exampleModalLabeledit'
+        aria-hidden='true'
+      >
+        <div className='modal-dialog'>
+          <div className='modal-content'>
+            <div className='modal-header'>
+              <h1 className='modal-title fs-5' id='exampleModalLabeledit'>
+              Edit Employee
+              </h1>
+              <button
+                type='button'
+                className='btn-close'
+                data-bs-dismiss='modal'
+                aria-label='Close'
+              ></button>
+            </div>
+            <div className='modal-body'>
+              <form id='addEmp'>
+                <div className='mb-3'>
+                  <label htmlFor='employee-name' className='col-form-label'>
+                    ID:
+                  </label>
+                  <input type='text' className='form-control' id='id' disabled/>
+                </div>
+                <div className='mb-3'>
+                  <label htmlFor='employee-name' className='col-form-label'>
+                    Name:
+                  </label>
+                  <input type='text' className='form-control' id='employee-name' />
+                </div>
+                <div className='mb-3'>
+                  <label htmlFor='role' className='col-form-label'>
+                    Role:
+                  </label>
+                  <input type='text' className='form-control' id='role' />
+                </div>
+                <div className='mb-3'>
+                  <label htmlFor='photo' className='col-form-label'>
+                    Image URL:
+                  </label>
+                  <input type='text' className='form-control' id='photo' />
+                </div>
+              </form>
+            </div>
+            <div className='modal-footer'>
+              <button type='button' className='btn btn-secondary' data-bs-dismiss='modal'>
+                Close
+              </button>
+              <button
+                type='submit'
+                className='btn btn-success'
+                data-bs-dismiss='modal'
+                onClick={handleeditsubmit}
+              >
+                Save
+              </button>
+            </div>
+          </div>
+        </div>
+      </div>
     </>
   );
 };
